Mandatory complete, except Makefile
The Makefile will need more research to do. The issue is the Makefile is relinking itself despite already being created. On the first try it works smoothly. However, when the second try was being made to test the makefile, it relinks by creating the file that already exists. Testing more test will be more needed to complete the project. There maybe some flaws to the program.